--- apps/openci_runner/pubspec_overrides.yaml
@@ -1,4 +1,0 @@
-# melos_managed_dependency_overrides: openci_models
-dependency_overrides:
-  openci_models:
-    path: ../../packages/openci_models
