+++ packages/openci_models/lib/src/converters/date_time_converter.dart
@@ -1,0 +1,13 @@
+import 'package:dart_firebase_admin/firestore.dart';
+import 'package:json_annotation/json_annotation.dart';
+
+class DateTimeConverter implements JsonConverter<DateTime, Timestamp> {
+  const DateTimeConverter();
+
+  @override
+  DateTime fromJson(Timestamp timestamp) =>
+      DateTime.fromMillisecondsSinceEpoch(timestamp.seconds * 1000);
+
+  @override
+  Timestamp toJson(DateTime dateTime) => Timestamp.fromDate(dateTime);
+}
