+++ packages/openci_models/lib/src/job/domain/command_log.dart
@@ -1,5 +1,6 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:openci_models/openci_models.dart';
+import 'package:openci_models/src/converters/date_time_converter.dart';
+import 'package:dart_firebase_admin/firestore.dart';
 
 part 'command_log.freezed.dart';
 part 'command_log.g.dart';
